requirements.txt
anthropic
pinecone
requests
python-dotenv